apps/zero/backend/flask/prod/requirements.txt
requests
flask
boto3
pyngrok
flask_socketio
gunicorn
gevent
flask_sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
